+++ src/main/java/com/j3d/engine/interact/cmd/commands/JaivaExecCmd.java
@@ -8,6 +8,7 @@ import com.j3d.engine.scene.nodes.geometry.GObject;
 import com.j3d.jaiva.TypeConverter;
 import com.j3d.ui.SafeJLabel;
 import com.j3d.utility.Parsing;
+import com.yetnt.utils.builders.AnsiColour;
 import com.yetnt.utils.builders.InlineHTML;
 import com.jaiva.errors.JaivaException;
 import com.jaiva.utils.generator.CCol;
@@ -39,8 +40,8 @@ public class JaivaExecCmd extends Command {
         } catch (Exception e) {
             if (e instanceof JaivaException j) {
                 logLabel.setText("Jaiva Execution Failed");
-                JOptionPane.showMessageDialog(StaticRefs.getMainFrame(), Parsing.removeCCol(j.getRawMessage()),  "Jaiva Execution Failed", JOptionPane.ERROR_MESSAGE);
-                StaticRefs.getLog().error("Jaiva Execution Failed: " + Parsing.removeCCol(j.getRawMessage()));
+                JOptionPane.showMessageDialog(StaticRefs.getMainFrame(), AnsiColour.remove(j.getRawMessage()),  "Jaiva Execution Failed", JOptionPane.ERROR_MESSAGE);
+                StaticRefs.getLog().error("Jaiva Execution Failed: " + AnsiColour.remove(j.getRawMessage()));
             } else {
                 throw new RuntimeException(e);
             }