+++ src/main/java/com/j3d/engine/react/history/History.java
@@ -75,7 +75,8 @@ public class History extends ArrayList<Action<?>> {
     }
 
     private static void repaint() {
-        StaticRefs.getMainPanel().repaint();
+        if (StaticRefs.getMainPanel() != null)
+            StaticRefs.getMainPanel().repaint();
     }
 
     // YOU ARE CHEATING.